Refactor ProvisionExtension feature for improved validation and handling
- Changed DeprovisionExtensionCommand to use int instead of decimal for Number.
- Updated DeprovisionExtensionCommandValidator to handle normalization exceptions.
- Refactored DeprovisionExtensionHandler to validate extension number and handle exceptions more clearly.
- Modified DeprovisionExtensionMappings to return a response indicating if the extension was removed.
- Enhanced DeprovisionExtensionResponse to include a Removed flag.
- Updated endpoint mappings to use int for route parameters.
- Improved ExtensionNumberNormalizer to handle int normalization.
- Adjusted tests for DeprovisionExtensionHandler and ProvisionExtensionHandler to reflect changes in command structure and validation.
- Cleaned up project references and ensured proper dependency management in csproj files.

+++ src/features/Extensions/ProvisionExtension/Deprovision/DeprovisionExtensionCommand.cs
@@ -2,4 +2,4 @@ using MediatR;
 
 namespace Zeus.Academia.Features.Extensions.ProvisionExtension.Deprovision;
 
-public sealed record DeprovisionExtensionCommand(decimal Number) : IRequest<DeprovisionExtensionResponse>;
+public sealed record DeprovisionExtensionCommand(int Number) : IRequest<DeprovisionExtensionResponse>;
